refactor: move identifier generation to schema
